requirements.txt
cma
argparse
gym[all]
box2d
tqdm
numpy
click
Pillow
pyyaml
test_tube
h5py
git+https://github.com/neuroevolution-ai/GymGuiEnvironments.git#egg=gym_gui_environments